appboxo/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-environ==0.4.5
django-redis==4.12.1
djangorestframework==3.12.2
pytz==2020.5
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
